refactor(#78): 로비 시작 검증 정책 분리
